added AnnouncementController + tests

+++ backend/src/main/java/dev/marco/example/springboot/service/impl/AnnouncementServiceImpl.java
@@ -1,34 +1,35 @@
 package dev.marco.example.springboot.service.impl;
 
-import org.apache.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import dev.marco.example.springboot.dao.impl.AnnouncementDAOImpl;
-import dev.marco.example.springboot.dao.impl.UserAnnouncementDAOImpl;
-import dev.marco.example.springboot.dao.impl.UserDAOImpl;
+import dev.marco.example.springboot.dao.AnnouncementDAO;
+import dev.marco.example.springboot.dao.UserAnnouncementDAO;
+import dev.marco.example.springboot.dao.UserDAO;
 import dev.marco.example.springboot.exception.*;
 import dev.marco.example.springboot.model.Announcement;
 import dev.marco.example.springboot.model.UserRoles;
 import dev.marco.example.springboot.service.AnnouncementService;
-
+import org.apache.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 
 import static dev.marco.example.springboot.exception.MessagesForException.*;
 
-
 @Service
 public class AnnouncementServiceImpl implements AnnouncementService {
 
     private static final Logger log = Logger.getLogger(AnnouncementServiceImpl.class);
+    private final AnnouncementDAO announcementDAO;
+    private final UserAnnouncementDAO userAnnouncementDAO;
+    private final UserDAO userDAO;
 
     @Autowired
-    AnnouncementDAOImpl announcementDAO;
-    @Autowired
-    UserAnnouncementDAOImpl userAnnouncementDAO;
-    @Autowired
-    UserDAOImpl userDAO;
+    private AnnouncementServiceImpl(AnnouncementDAO announcementDAO, UserAnnouncementDAO userAnnouncementDAO, UserDAO userDAO){
+        this.announcementDAO = announcementDAO;
+        this.userAnnouncementDAO = userAnnouncementDAO;
+        this.userDAO = userDAO;
+    }
 
     @Override
     public void setTestConnection() throws DAOConfigException {
@@ -52,7 +53,7 @@ public class AnnouncementServiceImpl implements AnnouncementService {
             throws UserException, AnnouncementException, DAOLogicException {
 
         try {
-            if(!userDAO.getUserById(announcement.getOwner()).getUserRole().equals(UserRoles.ADMIN))
+            if(!userDAO.getUserById(announcement.getIdUser()).getUserRole().equals(UserRoles.ADMIN))
                 throw new UserException(DONT_ENOUGH_RIGHTS);
 
             if(announcementDAO.isAnnouncementByTitle(announcement.getTitle())) {
@@ -67,13 +68,15 @@ public class AnnouncementServiceImpl implements AnnouncementService {
     }
 
     @Override
-    public void editAnnouncement(Announcement announcement, BigInteger idUser)
-            throws AnnouncementException, DAOLogicException {
+    public void editAnnouncement(Announcement announcement)
+            throws DAOLogicException, AnnouncementDoesNotExistException, UserDoesNotExistException, UserException {
         try {
-            if(!userDAO.getUserById(idUser).getUserRole().equals(UserRoles.ADMIN))
-                throw new AnnouncementException(DONT_ENOUGH_RIGHTS);
+            if(!announcementDAO.isAnnouncementById(announcement.getId()))
+                throw new AnnouncementDoesNotExistException(ANNOUNCEMENT_NOT_FOUND_EXCEPTION);
+            if(!userDAO.getUserById(announcement.getIdUser()).getUserRole().equals(UserRoles.ADMIN))
+                throw new UserException(DONT_ENOUGH_RIGHTS);
             announcementDAO.editAnnouncement(announcement);
-        } catch (DAOLogicException | UserDoesNotExistException e) {
+        } catch (DAOLogicException e) {
             log.error(DAO_LOGIC_EXCEPTION + MESSAGE_FOR_EDIT_ANNOUNCEMENT);
             throw new DAOLogicException(DAO_LOGIC_EXCEPTION, e);
         }
@@ -81,44 +84,56 @@ public class AnnouncementServiceImpl implements AnnouncementService {
 
     @Override
     public void deleteAnnouncement(BigInteger idAnnouncement, BigInteger idUser)
-            throws DAOLogicException, AnnouncementException {
+            throws DAOLogicException, UserException, AnnouncementDoesNotExistException, UserDoesNotExistException {
         try {
+            if(!announcementDAO.isAnnouncementById(idAnnouncement))
+                throw new AnnouncementDoesNotExistException(ANNOUNCEMENT_NOT_FOUND_EXCEPTION);
+
             if(!userDAO.getUserById(idUser).getUserRole().equals(UserRoles.ADMIN))
-                throw new AnnouncementException(DONT_ENOUGH_RIGHTS);
+                throw new UserException(DONT_ENOUGH_RIGHTS);
             announcementDAO.deleteAnnouncement(idAnnouncement);
-        } catch (UserDoesNotExistException e) {
+        } catch (DAOLogicException e) {
             log.error(DAO_LOGIC_EXCEPTION + MESSAGE_FOR_DELETE_ANNOUNCEMENT);
             throw new DAOLogicException(DAO_LOGIC_EXCEPTION, e);
         }
     }
 
     @Override
-    public void toLikeAnnouncement(BigInteger idAnnouncement, BigInteger idUser)
-            throws AnnouncementException, DAOLogicException {
+    public void setLikeAnnouncement(BigInteger idAnnouncement, BigInteger idUser)
+            throws DAOLogicException, AnnouncementDoesNotExistException, UserDoesNotExistException {
         try {
-            if(userAnnouncementDAO.isParticipant(idAnnouncement, idUser))
-                throw new AnnouncementException(ANNOUNCEMENT_ALREADY_LIKED);
-            userAnnouncementDAO.addParticipant(idAnnouncement, idUser);
-            announcementDAO.toLike(idAnnouncement);
+            if(!announcementDAO.isAnnouncementById(idAnnouncement))
+                throw new AnnouncementDoesNotExistException(ANNOUNCEMENT_NOT_FOUND_EXCEPTION);
+            userDAO.getUserById(idUser); // throw UserDoesNotExistException
+
+            if(userAnnouncementDAO.isParticipant(idAnnouncement, idUser)){
+                userAnnouncementDAO.deleteParticipant(idAnnouncement, idUser);
+                announcementDAO.toDisLike(idAnnouncement);
+            } else{
+                userAnnouncementDAO.addParticipant(idAnnouncement, idUser);
+                announcementDAO.toLike(idAnnouncement);
+            }
         } catch (DAOLogicException e) {
             log.error(DAO_LOGIC_EXCEPTION + MESSAGE_FOR_TO_LIKE_ANNOUNCEMENT);
             throw new DAOLogicException(DAO_LOGIC_EXCEPTION, e);
         }
     }
 
-    @Override
-    public void toDisLikeAnnouncement(BigInteger idAnnouncement, BigInteger idUser)
-            throws AnnouncementException, DAOLogicException {
-        try {
-            if(!userAnnouncementDAO.isParticipant(idAnnouncement, idUser))
-                throw new AnnouncementException(ANNOUNCEMENT_HAS_NOT_LIKED);
-            userAnnouncementDAO.deleteParticipant(idAnnouncement, idUser);
-            announcementDAO.toDisLike(idAnnouncement);
-        } catch (DAOLogicException e) {
-            log.error(DAO_LOGIC_EXCEPTION + MESSAGE_FOR_TO_DISLIKE_ANNOUNCEMENT);
-            throw new DAOLogicException(DAO_LOGIC_EXCEPTION, e);
-        }
-    }
+//    @Override
+//    public void toDisLikeAnnouncement(BigInteger idAnnouncement, BigInteger idUser)
+//            throws AnnouncementException, DAOLogicException, AnnouncementDoesNotExistException {
+//        try {
+//            if(!announcementDAO.isAnnouncementById(idAnnouncement))
+//                throw new AnnouncementDoesNotExistException(ANNOUNCEMENT_NOT_FOUND_EXCEPTION);
+//            if(!userAnnouncementDAO.isParticipant(idAnnouncement, idUser))
+//                throw new AnnouncementException(ANNOUNCEMENT_HAS_NOT_LIKED);
+//            userAnnouncementDAO.deleteParticipant(idAnnouncement, idUser);
+//            announcementDAO.toDisLike(idAnnouncement);
+//        } catch (DAOLogicException e) {
+//            log.error(DAO_LOGIC_EXCEPTION + MESSAGE_FOR_TO_DISLIKE_ANNOUNCEMENT);
+//            throw new DAOLogicException(DAO_LOGIC_EXCEPTION, e);
+//        }
+//    }
 
     @Override
     public List<Announcement> getPopularAnnouncements(int numberAnnouncements)
